Changed all float values to double and started implementation of statistic calculation

+++ src/main/java/com/example/levi9_challenge/dto/AdvancedStatsDTO.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AdvancedStatsDTO {
-    private float valorization;
-    private float effectiveFieldGoalPercentage;
-    private float trueShootingPercentage;
-    private float hollingerAssistRatio;
+    private double valorization;
+    private double effectiveFieldGoalPercentage;
+    private double trueShootingPercentage;
+    private double hollingerAssistRatio;
 }
